opt: Add correlated subquery Hibernate tests
Add several exemplar correlated subquery tests from the Hibernate
test suite, in order to ensure that we monitor their query plans going
forward. While several of the queries cannot be fully decorrelated,
upcoming enhancements will take care of that.

Release note: None
